adicao da relacao servico clinico e sector clinico
