+++ grails-app/domain/mz/org/fgh/sifmoz/backend/clinicSector/ClinicSector.groovy
@@ -7,6 +7,7 @@ class ClinicSector {
     String id
     String code
     String description
+   boolean active
 
     static belongsTo = [clinic: Clinic]
