+++ grails-app/domain/mz/org/fgh/sifmoz/backend/service/ClinicalService.groovy
@@ -1,6 +1,7 @@
 package mz.org.fgh.sifmoz.backend.service
 
 import com.fasterxml.jackson.annotation.JsonManagedReference
+import mz.org.fgh.sifmoz.backend.clinicSector.ClinicSector
 import mz.org.fgh.sifmoz.backend.episode.Episode
 import mz.org.fgh.sifmoz.backend.identifierType.IdentifierType
 import mz.org.fgh.sifmoz.backend.serviceattribute.ClinicalServiceAttribute
@@ -16,7 +17,8 @@ class ClinicalService {
 
     @JsonManagedReference
     static hasMany = [attributes: ClinicalServiceAttribute,
-                      therapeuticRegimens: TherapeuticRegimen]
+                      therapeuticRegimens: TherapeuticRegimen,
+                       clinicSectors: ClinicSector]
 
     static mapping = {
         id generator: "uuid"
@@ -26,6 +28,6 @@ class ClinicalService {
         description nullable: false
         attributes nullable: true
         therapeuticRegimens nullable: true
-
+        clinicSectors nullable: true
     }
 }
